Class extensions used to set public read-only properties rather than using builder objects.
The builder pattern does not seems a good fit for constructing immutable
value objects. Indeed it's implementation necessitated declaring instance
variables counterparts for public properties in order to set them, which
did not sit right with me. The implementation also produced a lot more code
than was desirable for simply setter properties on objects.

A new approach has been adopted that simply declares the setter methods
in a separate header file (...LJSSetters.h) and the readwrite properties
are defined in the value object's class implementation. Classes that construct
the (otherwise) immutable value objects import the LJSSetters.h header to
get access to the necessary property setters.

The main reason for this approach (and the pervious builder pattern approach)
is to provide a set of value objects the API returns that are immutable.
However they must be constructed in the first place by the API. So the API
constructs LJSStop, LJSService, and LJSDeparture objects but clients
of the API are only be able to consume them.