Raw departure date string saved to Departure model.
HTML data represents the departure dates as either times "11:12" or pretty
user friendly strings "2 Mins". This pretty string is saved along with a
parsed NSDate instance for the departure date. The pretty string will be
used in the UI, the NSDate used in data manipulation like ordering departures
in chronological order.