Completion block nilled out only when another completion block has not been set.
Only nil out the completion block if it is the same block
as the one that was set before it was called.
If the completion block calls liveDataAtURL:completion:
the another completion block will be set before we have
change to nil it out. If we nil out the completion block
in this case then the second liveDataAtURL:completion: call
will not get a completion block call back.