Amend client to return a Result object for live data
A LJSLiveDataResult encapsulates all the live data the client can return:
A Stop object representing the successful parsing of departure data.
An array of messages represents the successful parsing of any messages in
the HTMl.

By wrapping the Stop and the messages up in a Result object it is clearer
that the client can finish successfully without a Stop (when there are no
departures for the next hour) but with still have some messages. Conversely
the client can finish successfully with a Stop and no messages. In both
of these successful scenarios you will be returned a Result object.